+++ src/DotNext/Span.cs
@@ -39,7 +39,7 @@ public static partial class Span
         if (span.IsEmpty)
             return salted ? RandomExtensions.BitwiseHashSalt : 0;
 
-        return Intrinsics.GetHashCode32Unaligned(ref As<T, byte>(ref MemoryMarshal.GetReference(span)), checked((nint)span.Length * sizeof(T)), salted);
+        return Intrinsics.GetHashCode32Unaligned(ref As<T, byte>(ref MemoryMarshal.GetReference(span)), checked((nuint)span.Length * (nuint)sizeof(T)), salted);
     }
 
     /// <summary>
@@ -101,7 +101,7 @@ public static partial class Span
         where THashFunction : struct, IConsumer<int>
     {
         if (!span.IsEmpty)
-            Intrinsics.GetHashCode32Unaligned(ref hashFunction, ref As<T, byte>(ref MemoryMarshal.GetReference(span)), checked((nint)span.Length * sizeof(T)));
+            Intrinsics.GetHashCode32Unaligned(ref hashFunction, ref As<T, byte>(ref MemoryMarshal.GetReference(span)), checked((nuint)span.Length * (nuint)sizeof(T)));
 
         if (salted)
             hashFunction.Invoke(RandomExtensions.BitwiseHashSalt);
@@ -147,7 +147,7 @@ public static partial class Span
         where THashFunction : struct, IConsumer<long>
     {
         if (!span.IsEmpty)
-            Intrinsics.GetHashCode64Unaligned(ref hashFunction, ref As<T, byte>(ref MemoryMarshal.GetReference(span)), checked((nint)span.Length * sizeof(T)));
+            Intrinsics.GetHashCode64Unaligned(ref hashFunction, ref As<T, byte>(ref MemoryMarshal.GetReference(span)), checked((nuint)span.Length * (nuint)sizeof(T)));
 
         if (salted)
             hashFunction.Invoke(RandomExtensions.BitwiseHashSalt);
@@ -212,7 +212,7 @@ public static partial class Span
         if (span.IsEmpty)
             return salted ? RandomExtensions.BitwiseHashSalt : 0L;
 
-        return Intrinsics.GetHashCode64Unaligned(ref As<T, byte>(ref MemoryMarshal.GetReference(span)), checked((nint)span.Length * sizeof(T)), salted);
+        return Intrinsics.GetHashCode64Unaligned(ref As<T, byte>(ref MemoryMarshal.GetReference(span)), checked((nuint)span.Length * (nuint)sizeof(T)), salted);
     }
 
     /// <summary>
